Moved dasmooi commands to seperate cog and organised them slightly
